Automatically detect the input file type
Instead of assuming that the (first) input file is a JPEG file by
default, be smart and automatically detect the file type by reading the
headers. Users can still force the file to be interpreted as a PPM or as
a JPEG by using --ppm or the new --jpeg command line arguments.

Signed-off-by: Lukas Fleischer <lfleischer@lfos.de>